requirements.txt
chainlit==1.0.0
pinecone-client==3.0.0
sentence-transformers==2.2.2
google-generativeai==0.3.2
langgraph==0.2.16
langchain-core==0.3.15
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0